fixup! feat(ng-dev): enable the use of ng-dev credential service for pr merging
